+++ src/UseCases/GenerateIban/GenerateIbanUseCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WMDE\Fundraising\Frontend\UseCases\ValidateBankData;
+namespace WMDE\Fundraising\Frontend\UseCases\GenerateIban;
 
 use WMDE\Fundraising\Frontend\BankDataConverter;
 
@@ -8,14 +8,10 @@ use WMDE\Fundraising\Frontend\BankDataConverter;
  * @licence GNU GPL v2+
  * @author Kai Nissen <kai.nissen@wikimedia.de>
  */
-class ValidateBankDataUseCase {
+class GenerateIbanUseCase {
 
 	public function __construct( BankDataConverter $bankDataConverter ) {
 		$this->bankDataConverter = $bankDataConverter;
-	}
-
-	public function checkIban( string $iban ) {
-		return $this->bankDataConverter->getBankDataFromIban( $iban );
 	}
 
 	public function generateIban( string $accountNumber, string $bankCode ) {
